board functions all moved
