+++ app/main.py
@@ -1,17 +1,12 @@
 import json
 import os
 import bottle
-from parseBoard import setBoardValues
+import parseBoard
+from parseBoard import printBoard
 
 from api import ping_response, start_response, move_response, end_response
 from pathfinding.core.grid import Grid
 from pathfinding.finder.a_star import AStarFinder
-
-SNAKE = -1
-WALL_SPACE = 10
-OPEN_SPACE = 5
-SMALLER_SNAKE_FUTURE_HEAD = 3
-ourHead = 1
 
 
 @bottle.route('/')
@@ -65,15 +60,16 @@ def move():
     # Converts data to be parsable
     converted_data = json.loads(json.dumps(data))
 
-    board = boardToArray(converted_data)
+    board = parseBoard.boardToArray(converted_data)
 
     # closeFood = getNearestFood(converted_data)
-    pathable_board = setBoardValues(converted_data)
+    pathable_board = parseBoard.setBoardValues(converted_data)
 
     # Json data is printed for debug help
     print(json.dumps(data))
     printBoard(board, pathable_board)
-    directions = cardinal(converted_data, getMinPathToFood(converted_data, pathable_board))
+    directions = cardinal(converted_data, getMinPathToFood(converted_data, 
+                                                           pathable_board))
 
     direction = directions[0]
 
@@ -91,57 +87,6 @@ def end():
     print(json.dumps(data))
 
     return end_response()
-
-
-def boardToArray(dataDump):
-    """
-    Converts converted JSON data from battlesnake engine to a list
-    Args:
-        dataDump (list): Converted JSON data
-
-    Returns:
-        2D list of game board.
-    """
-    board_width = dataDump["board"]["width"]
-    board_height = dataDump["board"]["height"]
-    board = [[0 for x in range(board_width)] for y in range(board_height)]
-    # label spaces as food
-    for z in dataDump["board"]["food"]:
-        x = z['x']
-        y = z['y']
-        board[y][x] = 'F'
-    # finding your body
-    me = dataDump["you"]["id"]
-    for z in dataDump["you"]["body"]:
-
-        if (z == dataDump["you"]["body"][0]):
-            x = z['x']
-            y = z['y']
-            board[y][x] = 'H'
-        else:
-            x = z['x']
-            y = z['y']
-            board[y][x] = 'S'
-    # to find other snakes
-    for z in dataDump["board"]["snakes"]:
-        name = z["id"]
-        for a in z["body"]:
-            if (name != me):
-                if (a == z["body"][0]):
-                    x = a['x']
-                    y = a['y']
-                    board[y][x] = 'E'
-                else:
-                    x = a['x']
-                    y = a['y']
-                    board[y][x] = 'S'
-    return board
-
-
-
-
-
-
 
 
 def getMinPathToFood(converted_data, pathBoard):
@@ -232,54 +177,6 @@ def cardinal(converted_data, path):
         else:
             direction = ['left']
     return direction
-
-
-def bullyPathing(converted_data, pathBoard):
-    """
-    Finds Snakes that are smaller than us and assigns the area around their head with the Smaller_Snake_Future_Head
-
-    May need logical update for snake bodies but it should be fine
-
-    Args:
-        converted_data (json): python readable json
-        path (list): path from a*
-
-    Return:
-        updated board (list) with potentially new values around smaller snakes heads
-    """
-    me = converted_data["you"]["id"]
-    board = pathBoard
-    # other snakes heads will be assigned xy
-    for z in converted_data["board"]["snakes"]:
-        name = z["id"]
-        for a in z["body"]:
-            if ((str(name) != str(me)) and (len(z["body"]) < len(converted_data["you"]["body"]))):
-                if (a == z["body"][0]):
-                    x = a['x']
-                    y = a['y']
-                    if(x < converted_data['board']['width'] and x > 0):
-                        if(pathBoard[y][x+1] != -1):
-                            board[y][x+1] = 3
-                        if(pathBoard[y][x-1] != -1):
-                            board[y][x-1] = 3
-                    if(y < converted_data['board']['width'] and y > 0):
-                        if(pathBoard[y+1][x] != -1):
-                            board[y+1][x] = 3
-                        if(pathBoard[y-1][x] != -1):
-                            board[y-1][x] = 3
-
-    return board
-
-
-def printBoard(converted_board, integer_board):
-    for x in integer_board:
-        for y in x:
-            print(str(y) + " "),
-        print()
-    for x in converted_board:
-        for y in x:
-            print(str(y) + " "),
-        print()
 
 
 # Expose WSGI app (so gunicorn can find it)
